requirements.txt
pystray
pillow
pyobjc-framework-Cocoa; sys_platform == "darwin"
